+++ src/main/resources/db/migration/changes/Users.json
@@ -109,7 +109,7 @@
                     "constraints": {
                       "nullable": true
                     },
-                    "name": "MODERATION",
+                    "name": "MODERATOR",
                     "type": "BOOLEAN"
                   }
                 }]
